renamed to create_vocabulary
